Create profile Icon component.
