+++ frontend/src/components/Authenticated.jsx
@@ -2,13 +2,13 @@ import {
     Route, Switch,
 } from "react-router-dom";
 import React from "react";
-import {Account} from "../pages/Account";
+import {AccountPage} from "../pages/AccountPage";
 import {NotFound} from "../pages/NotFound";
 
 export const Authenticated = () => {
     return (
         <Switch>
-            <Route path="/account" component={Account} />
+            <Route path="/account" component={AccountPage} />
             <Route component={NotFound} />
         </Switch>
     )
