+++ frontend/src/components/css/ProfileIcon.css
@@ -1,0 +1,30 @@
+.profile-div {
+    margin-left: auto;
+    display: flex;
+}
+
+.profile-username {
+    font-weight: 300;
+    margin-right: 20px;
+}
+.profile {
+    margin-left: auto;
+    margin-right: 1vw;
+    max-width: 40px;
+    transition: transform 0.3s ease-in-out;
+}
+
+.profile:hover {
+    transform: scale(1.15);
+}
+
+@media screen and (max-width: 550px) {
+    .profile-username {
+        font-size: 18px;
+        margin-right: 10px;
+    }
+
+    .profile {
+        max-width: 27px;
+    }
+}
