+++ frontend/src/pages/css/RoomSelect.css
@@ -15,17 +15,6 @@ header img.logo {
     max-width: 70px;
 }
 
-header img.profile {
-    margin-left: auto;
-    margin-right: 1vw;
-    max-width: 40px;
-    transition: transform 0.3s ease-in-out;
-}
-
-header img.profile:hover {
-    transform: scale(1.15);
-}
-
 /*
 Styling for Room Form
 */
@@ -173,10 +162,6 @@ Styling for Error Message
         max-width: 60px;
     }
 
-    header img.profile {
-        max-width: 35px;
-    }
-
     form.roomCode input {
         font-size: 14px;
     }
@@ -193,11 +178,6 @@ Styling for Error Message
 
     header img.logo {
         max-width: 45px;
-    }
-
-    header img.profile {
-        padding-right: 5px;
-        max-width: 30px;
     }
 
     form.roomCode hr {
